[#90] 다이나믹 사이즈 매니저 구조체로 변경
